requirements.txt
torch
torchaudio
torchvision
torchtext
numpy
librosa
scipy
tensorboard
soundfile
matplotlib
numba
tgt
pandas
pytorch-lightning==1.4.9
pyyaml